+++ ov_core/src/feat/Feature.h
@@ -53,13 +53,13 @@ namespace ov_core {
         std::unordered_map<size_t, std::vector<Eigen::VectorXf>> uvs_norm;
 
         /// Timestamps of each UV measurement (mapped by camera ID)
-        std::unordered_map<size_t, std::vector<float>> timestamps;
+        std::unordered_map<size_t, std::vector<double>> timestamps;
 
         /// What camera ID our pose is anchored in!! By default the first measurement is the anchor.
         int anchor_cam_id = -1;
 
         /// Timestamp of anchor clone
-        float anchor_clone_timestamp;
+        double anchor_clone_timestamp;
 
         /// Triangulated position of this feature, in the anchor frame
         Eigen::Vector3f p_FinA;
@@ -76,7 +76,7 @@ namespace ov_core {
          *
          * @param valid_times Vector of timestamps that our measurements must occur at
          */
-        void clean_old_measurements(std::vector<float> valid_times);
+        void clean_old_measurements(std::vector<double> valid_times);
 
         /**
          * @brief Remove measurements that are older then the specified timestamp.
@@ -85,7 +85,7 @@ namespace ov_core {
          *
          * @param timestamp Timestamps that our measurements must occur after
          */
-        void clean_older_measurements(float timestamp);
+        void clean_older_measurements(double timestamp);
 
     };
