+++ ov_eval/src/calc/ResultTrajectory.h
@@ -157,7 +157,7 @@ namespace ov_eval {
     protected:
 
         // Trajectory data (loaded from file and timestamp intersected)
-        std::vector<float> est_times, gt_times;
+        std::vector<double> est_times, gt_times;
         std::vector<Eigen::Matrix<float,7,1>> est_poses, gt_poses;
         std::vector<Eigen::Matrix3f> est_covori, est_covpos, gt_covori, gt_covpos;
